found error when deleting wrecks (delete SE* then trying to bcast = segfault) fixed that segfault and adjusted other SE->Delete() code to avoid potential segfaults few other misc code updates (improvements)

+++ src/eve-server/EVEServerConfig.h
@@ -205,7 +205,7 @@ public:
         bool ssl;
         bool useSocket;
         bool autoReconnect;
-        uint8 dbTimeout;
+        uint dbTimeout;
     } database;
 
     // From <files/>
